Rename app user to users, change path api/users/register > api/users/
