+++ library_service/urls.py
@@ -20,5 +20,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("api/users/", include("user.urls", namespace="user")),
+    path("api/users/", include("users.urls", namespace="users")),
 ]
