--- user/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from user.views import CreateUserView
-
-app_name = "user"
-
-urlpatterns = [
-    path("register/", CreateUserView.as_view(), name="register")
-]
