+++ users/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class UserConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "user"
+    name = "users"
