+++ users/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from users.views import CreateUserView
+
+app_name = "users"
+
+urlpatterns = [
+    path("", CreateUserView.as_view(), name="register")
+]
