+++ users/views.py
@@ -1,6 +1,6 @@
 from rest_framework import generics
 
-from user.serializers import UserSerializer
+from users.serializers import UserSerializer
 
 
 class CreateUserView(generics.CreateAPIView):
